move remaining objs from allocator_types to allocator
