+++ tt_metal/impl/allocator/algorithms/allocator_algorithm.hpp
@@ -10,7 +10,7 @@
 
 #include <tt_stl/assert.hpp>
 #include <hal_types.hpp>
-#include <allocator_types.hpp>
+#include <tt-metalium/allocator.hpp>
 
 namespace tt {
